test: add unit tests for StockProfitService to validate profit calculations

+++ src/main/java/org/example/stockcalculator/repository/UserAccountRepository.java
@@ -3,7 +3,7 @@ package org.example.stockcalculator.repository;
 import org.example.stockcalculator.entity.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserAccountRepository extends JpaRepository<UserAccount, Integer > {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Long > {
 
     UserAccount findByEmailIgnoreCase(String email);
 }
